v3.87a - minor warning fixes for use under MSVC

+++ SG14_test/plf_colony_test_suite.cpp
@@ -787,7 +787,7 @@ void plf_colony_test_suite()
 				{
 					it2 = it1 = i_colony.begin();
 
-					size = i_colony.size();
+					size = static_cast<unsigned int>(i_colony.size());
 					range1 = xor_rand() % size;
 					range2 = range1 + 1 + (xor_rand() % (size - range1));
 					i_colony.advance(it1, range1);
